--- pages/cart/config.js
@@ -1,29 +1,0 @@
-export const STEPS = [
-  {
-    label: 'Cart',
-    isComplete: false
-  },
-  {
-    label: 'Delivery',
-    isComplete: false
-  },
-  {
-    label: 'Payment',
-    isComplete: false
-  },
-  {
-    label: 'Finish',
-    isComplete: false
-  }
-]
-
-export const DELIVERY_OPTIONS = [
-  {
-    value: 'delivery',
-    label: 'Delivery'
-  },
-  {
-    value: 'pickup',
-    label: 'Pickup'
-  }
-]
